+++ src/pantr/bspline/_bspline.py
